GPU Particleの更新処理を追加

+++ Engine/Contexts/RenderContext/RenderContext.cpp
@@ -8,26 +8,26 @@ RenderContext::RenderContext(Renderer* renderer, LightManager* lightManager) {
 }
 
 void RenderContext::DrawModel(Model* model, BlendMode blendMode) {
-	renderer_->DrawModel(model, lightManager_, static_cast<int>(blendMode));
+	renderer_->DrawModel(model, lightManager_, blendMode);
 }
 
 void RenderContext::DrawSprite(Sprite* sprite, BlendMode blendMode) {
-	renderer_->DrawSprite(sprite, static_cast<int>(blendMode));
+	renderer_->DrawSprite(sprite, blendMode);
 }
 
 void RenderContext::DrawInstancedModel(InstancedModel* instancedModel, BlendMode blendMode) {
-	renderer_->DrawModelInstance(instancedModel, lightManager_, static_cast<int>(blendMode));
+	renderer_->DrawModelInstance(instancedModel, lightManager_, blendMode);
 }
 
 void RenderContext::DrawParticle(ParticleSystem* particleSystem, BlendMode blendMode) {
-	renderer_->DrawParticles(particleSystem, static_cast<int>(blendMode));
+	renderer_->DrawParticles(particleSystem, blendMode);
 }
 void RenderContext::DrawGPUParticle(ParticleSystem* particleSystem, BlendMode blendMode) {
-	renderer_->DrawGPUParticle(particleSystem, static_cast<int>(blendMode));
+	renderer_->DrawGPUParticle(particleSystem, blendMode);
 }
 
 void RenderContext::DrawPrimitive(Primitive* primitive, BlendMode blendMode) {
-	renderer_->DrawPrimitive(primitive, static_cast<int>(blendMode));
+	renderer_->DrawPrimitive(primitive, blendMode);
 }
 
 void RenderContext::DrawSkybox(Texture* skybox) {
